+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-03 14:26:12
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-07-03 19:47:51
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '2532';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-07-03T14:26:12+05:00');
+  final String commitRef = '2549';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-07-03T19:47:51+05:00');
 }